Squashed commit of the following:
commit 1cd9cee0c8e3f0771915707ff6cd05fdcabc4d6b
Author: Gary Tempus <gtempus@users.noreply.github.com>
Date:   Fri Jul 11 13:20:27 2025 -0400

    PZB-169: Refactor to Use Background Task for Zonal Stats and Improve FastAPI Organization (#25)

    * feat: run analytics in a fastapi `BackgroundTask` after creating the resource

    * feat: return info for pending analytics in the `get` response

    * :fire: refactor: remove unused routes

    * refactor: reorganize router configuration

    * refactor: reorganize the route models

    * refactor: reorganize api test module

    * :bug: fix(Dist Alerts): metadata for a pending request was not being reloaded

commit 6bed2a83fb4f582e6c699becad5350ce11a52c86
Author: jterry64 <justin.terry@wri.org>
Date:   Thu Jul 10 14:44:55 2025 -0700

    Add terraform for API service to deploy on ECS with a load balancer (using existing VPC)

commit 6d80c76dfbf8e2d301e6d705a4f811823417b925
Author: danscales <dan.scales@gmail.com>
Date:   Mon Jul 7 14:46:17 2025 -0700

    Switch to non-clipped versions of datasets (#24)

    - Use the non-clipped versions of datasets and do xr.align on the fly.
     - Moved a bunch of duplicated constants to pipelines/globals.py.
     - Made some names consistent (singular rather than plural)

    Still need to create a non-clipped zarr of natural lands.
