+++ api/app/routers/land_change/dist_alerts/__init__.py
@@ -1,0 +1,66 @@
+from .dist_alerts import create
+
+create.__doc__ = """
+    # Primary Dataset
+    | Dataset Key Aspects | Detailed Description |
+    |---------------------|----------------------|
+    | **Title** | Global all ecosystem disturbance alerts (DIST-ALERT) |
+    | **Description** <br> A descriptive summary of what the dataset is and key details, such as information on geographic coverage, temporal range, and the data sources used. | Monitors global vegetation disturbance in near-real-time using harmonized Landsat-Sentinel-2 (HLS) imagery |
+    | **Function / usage notes in Zeno** <br> The intended use of the dataset. How might users employ this data in order to have some impact in the world?<br>When should Zeno suggest/use this dataset? Please include examples of use cases (e.g. X dataset can be used for XYZ). You can also include these examples in the prompt spreadsheet. | Users are informed about significant changes to the vegetation, which can indicate human induced changes such as deforestation, as well as drought and other events where actions by land managers may be required. Crucially, this dataset covers all vegetation, not just forest, so that new non-conversion of natural ecosystems commitments can be monitored. |
+    | **Methodology** <br> How was this dataset created? | This dataset is a derivative of the OPERA’s DIST-ALERT product (OPERA Land Surface Disturbance Alert from Harmonized Landsat Sentinel-2 product), which is derived through a time-series analysis of harmonized data from the NASA/USGS Landsat and ESA Sentinel-2 satellites (known as the HLS dataset). The product identifies and continuously monitors vegetation cover change in 30-m pixels across the globe. It can be accessed through the LPDAAC website here: Earthdata Search , and on Google Earth Engine (GEE) as an Image Collection with asset ID: projects/glad/HLSDIST/current/[Insert Layer Name].<br>The DIST-ALERT on GFW is a derivative of this, and additional data layers not used in the GFW product are available through the LPDAAC and GEE such as initial vegetation fraction, and disturbance duration. While the version on the LPDAAC is updated every 2-4 days, the data is updated weekly on GFW.  The product detects notable reductions in vegetation cover (measured as “vegetation fraction” or the percent of the ground that is covered by vegetation) for every pixel every time the new satellite data is acquired and the ground is not obscured by clouds or snow.  <br>The current vegetation fraction estimate is compared to the minimum fraction for the same time period (within 15 days before and after) in the previous 3 years, and if there is a reduction, then the system identifies an alert in that pixel. Anomalies of at least a 10% reduction from the minimum become alerts in the original product, and on GFW, a higher threshold of 30% is used, to reduce noise, and false alerts in the dataset. Because the product compares each pixel to the minimum for the same time period in previous years, it takes into account regular seasonal variation in vegetation cover. As the product is global and detects vegetation anomalies, much of the data may not be applicable to GFW users monitoring forests. Therefore, we also offer the option to mask the alerts with UMD’s tree cover map, allowing users to view only alerts within 30% canopy cover. |
+    | **Cautions in Zeno**<br>What should be kept in mind when using this dataset?<br>When should Zeno not suggest/use this dataset? Feel free to include examples of when the data should not be used (e.g. X dataset cannot be used for XYZ).<br>Generates a persistent resource link for deforestation alerts analytics requests. | These alerts detect vegetation cover loss or disturbance. This product does not distinguish between human-caused and other disturbance types. For example, where alerts are detected within plantation forests, alerts may indicate timber harvesting operations, without a conversion to a non-forest land use, and when alerts are detected within crop land, alerts may represent crop harvesting<br>We do not recommend using the alerts for global or regional trend assessment, nor for area estimates. Rather, we recommend using the annual tree cover loss data for a more accurate comparison of the trends in forest change over time, and for area estimates. For global land cover changes, 20 year data may be more useful for certain purposes and should be explored. Additionally, updates to the methodologies and variation in cloud cover between months and years pose additional risks to using alerts for inter/intra-annual comparison.<br>The alerts can be ‘curated’ to identify those alerts of interest to a user, such as those alerts which are likely to be deforestation and might be prioritized for action. A user can do this by overlaying other contextual datasets, such as protected areas, or planted trees. The non-curated data are provided here in order that users can define their own prioritization approaches. Curated alert locations within tree cover are provided in the Places to Watch data layer.<br>We provide a masked version of the product within “tree cover” which is defined as all vegetation greater than 5 meters in height (2020) with greater than 30% canopy cover (2010), and may take the form of natural forests or plantations. Annual tree cover loss from 2021 is masked out. <br>In contrast to other alert systems available on GFW, DIST-ALERT continues to monitor pixels where it has identified an alert in the past. The DIST-ALERT retains the date of the most recent detection of disturbance, keeping users informed of the most up-to-date changes within tree cover.<br>Two confidence levels are provided. The approach determines confidence level by the number of anomalous observations, with more observations meaning a higher confidence level. That is, two to three anomalies detected result in a low confidence alert, whereas four or more mean a high confidence alert. UMD’s Google Earth Engine app DIST-ALERT ) displays alerts, with a different approach used to define confidence, and a different threshold for the vegetation reduction which triggers alerts. |
+    | **Resolution** | 30 x 30 meters |
+    | **Geographic Coverage** | Global |
+    | **Update Frequency** | Underlying product updated daily, with image revisit time every 2-4 days. Product on GFW updated weekly. |
+    | **Content Date** | 1 January 2023 – present (GFW has data since 1 December 2023, and in future, will display only the most recent 2 years of alert data) |
+    | **Keywords** | land cover, alerts, deforestation, conversion, vegetation disturbance |
+    | **Band values/meaning**<br>What does the pixel values of each band mean? What are the possible values? | The input layers and their encoding can be found in the documentation<br>Single bands<br>Type: discrete<br>default.tif (formula is (10000 * confidence) + (days since Dec 31, 2020))<br>20000-29999: Low confidence (Ex. 21431 = Low confidence alert on Dec 1, 2024)<br>30000-39999: High confidence (Ex. 31433 = High confidence alert on Dec 3, 2024)<br>intensity.tif<br>All alerts set to 255 and use bilinear resampling on the overviews so that areas with dense alerts have higher values<br>Possible values of the final GFW product:<br>either low or high confidence, based on the alert count<br>date of the alert based on the VEG-DIST-DATE (this represents the Day of first loss anomaly detection of the most recent disturbance event. Day denoted as the number of days since December 31, 2020.) |
+
+    -----
+    # Available Intersection Datasets
+    ## Drivers (LDACS)
+    | Dataset Key Aspects | Detailed Description |
+    |---------------------|----------------------|
+    | **Title** | Land Disturbance Alert Classification System (LDACS) |
+    | **Description** <br>A descriptive summary of what the dataset is and key details, such as information on geographic coverage, temporal range, and the data sources used. | The LDACS classifies the drivers of global vegetation disturbance alerts (DIST-ALERT)(Pickens, Hansen, and Song, 2024) on a quarterly basis using a rule-based approach. Alerts are categorized as potential conversion, wildfires, surface water dynamics, or crop management. This system integrates DIST-ALERT data with thematic layers on fire activity (NASA VIIRS Fire Alerts), surface water dynamics (Pickens et al. 2020), cropland extent (UMD Global Cropland Dynamics (Potapov et al. 2022); ESA WorldCover (Zanaga et al. 2020)), and natural land cover (SBTN Natural Lands), using a rule-based approach to assign probable causes to alerts. Each run of the system processes DIST-ALERT data from the previous 12 months and is updated quarterly. |
+    | **Function / usage notes in Zeno**<br>The intended use of the dataset. How might users employ this data in order to have some impact in the world?<br>When should Zeno suggest/use this dataset? Please include examples of use cases (e.g. X dataset can be used for XYZ). You can also include these examples in the prompt spreadsheet. | The LDACS is designed to help users distinguish between disturbances caused by human activity and those resulting from natural events. DIST-ALERTs are assigned to one of five categories: potential conversion, crop management, wildfire, surface water dynamics, or unclassified. The LDACS simplifies the process of identifying alerts that are likely caused by land conversion, with the goal of supporting broader adoption of DIST-ALERT among governments, companies, and conservation organizations. <br>Zeno should use this dataset when a user is interested in why a vegetation disturbance alert has occurred. The LDACS dataset has global coverage and therefore, can be used across all vegetated ecosystems. LDACS data will be made available for the most recent 12-month period and will be updated quarterly. |
+    | **Methodology**<br>How was this dataset created? | The Land Disturbance Alert Classification System was developed by combining DIST-ALERT data with a set of thematic data masks derived from multiple geospatial data sources. These masks were used to infer the most likely driver of each vegetation disturbance alert. <br>Each run of the system processes DIST-ALERT data from the previous 12 months, using the alert start date (VEG-DIST-DATE) to filter the data, to ensure that classifications are based on recent and relevant disturbance activity. Because each run processes a rolling 12-month window, a single disturbance event may appear in multiple consecutive quarterly runs (up to three or four times), depending on the timing of its detection. Only DIST-ALERTS that fall under the following criteria were classified under this system: VEG-DIST-COUNT >= 2 and VEG-ANOM-MAX > 50, meaning alerts with less than 50% vegetation loss or fewer than two cloud-free satellite observations are excluded to minimize false positives. <br>The following paragraphs describe the classification logic for each disturbance category:<br>Wildfire: DIST-ALERTs were flagged as fire-related if they occurred within the three months after, and within 3 kilometers of, a VIIRS fire alert, and were located within natural lands as defined by the SBTN Natural Lands Map.<br>Flooding: To identify alert pixels likely associated with surface water dynamics, we used annual surface water dynamic composites (Pickens et al. 2020) to create a binary mask highlighting all pixels where surface water was detected at least once between 1999 and 2021. This mask identifies areas with a history of surface water presence and helps flag locations prone to periodic or irregular inundation. Alerts that fall within the surface water dynamics mask are flagged and classified as flooding. <br>Crop management: DIST-ALERT pixels located within areas mapped as cropland by either the UMD Global Cropland Dynamics dataset (Potapov et al. 2022) or the ESA WorldCover dataset (Zanaga et al. 2020) were classified as crop management changes.<br>Potential conversion: Potential conversion events were then assessed using two metrics from the DIST-ALERT data set: maximum vegetation anomaly (VEG-ANOM-MAX) and alert duration (VEG-DIST-DUR). Alerts were flagged as potential conversion if they occurred in natural lands (SBTN Natural Lands), showed a vegetation loss greater than 50% at any point during the alert period (VEG-ANOM-MAX > 50), and that loss persisted for at least 75 days (VEG-DIST-DUR >= 75). The potential conversion alert pixels were post-processed using morphological operations, a common technique in satellite image analysis to clean up pixel-based maps (Haralick, Sternberg, and Zhuang 1987; Löw et al. 2022; d’Andrimont et al. 2020). Specifically, erosion was applied to remove small or narrow clusters of pixels that may represent noise or false detections. Dilation was then applied to expand the remaining patches outward by one pixel in all directions. While this step helps restore the general shape and spatial extent of larger disturbance areas, it produces a smoothed approximation of the original patch, helping eliminate small gaps and jagged edges while maintaining the overall structure.<br>The resulting classes are merged into a final classification layer using a hierarchical scheme: wildfire takes the highest priority, followed by flooding, crop management, and potential conversion.<br>Unclassified: Alerts that do not meet the criteria for any of the four categories above are assigned to a fifth class, labeled “unclassified” alerts. |
+    | **Cautions in Zeno**<br>What should be kept in mind when using this dataset?<br>When should Zeno not suggest/use this dataset? Feel free to include examples of when the data should not be used (e.g. X dataset cannot be used for XYZ). | In contrast to other alert systems, DIST-ALERT continues to monitor pixels where it has identified an alert in the past. The DIST-ALERT retains the date of the most recent detection of disturbance, keeping users informed of the most up-to-date changes.<br>Several input datasets used to produce the LDACS contain temporal gaps or known mapping errors that may affect classification accuracy. Please refer to the technical note for more information on the limitations of the input datasets.<br>The classification system uses a static version of the SBTN natural lands map to identify potential conversion events. This layer is not updated in response to disturbances flagged by DIST-ALERT. For example, if a wildfire occurs in a natural area and is followed by land clearing, the system will still treat the area as natural land if it remains within the boundaries of the original natural lands data set. Updates to this map only occur when the source data set itself is revised, which may result in alerts being evaluated against outdated land use information. As a result, disturbances that alter the ecological state of the land (e.g., severe fire or tree cover loss) are not reflected in the natural lands mask used for classification, potentially contributing to inclusion or exclusion errors depending on the disturbance sequence and timing.<br>Because potential conversion is identified based on the duration of vegetation loss, there is an inherent delay in labeling these alerts. Although vegetation loss may be detected shortly after clearing, alerts are not assigned to the potential conversion class until they exceed the 75-day duration threshold. This delay allows for greater confidence in classification but may reduce the timeliness of conversion detection for rapid response. <br>LDACS data will be made available for the most recent 12-month period and will be updated quarterly. Data prior to this 12-month period will not be available. Additionally, there may be up to a 3-month delay in LDACS data until the quarterly update occurs.<br>LDACS classifies DIST-ALERTs with a minimum of VEG-DIST-COUNT >= 2 and VEG-ANOM-MAX > 50. The DIST-ALERT layer on Zeno has a minimum of VEG-DIST-COUNT >-2 and VEG-ANOM-MAX >= 30, therefore, LDACS classifies a subset of these alerts. DIST-ALERTs with a VEG-ANOM-MAX between 30 to 50 will not be classified through the LDACS. |
+    | **Resolution** | 30 x 30 meters |
+    | **Geographic Coverage** | Global |
+    | **Update Frequency** | Quarterly |
+    | **Content Date** | Alerts evaluated over previous 12-month period, updated quarterly. |
+    | **Keywords** | alerts, deforestation, conversion, vegetation disturbance, drivers of disturbance |
+    | **Band values/meaning**<br>What does the pixel values of each band mean? What are the possible values? | Single band<br>Type: classification<br>Possible values:<br>1 = Wildfire<br>2 = Flooding<br>3 = Crop management<br>4 = Potential conversion<br>5 = Unclassified |
+
+    -----
+
+    ## Natural Lands (SBTN)
+    | Dataset Key Aspects | Detailed Description |
+    |---------------------|----------------------|
+    | **Title** | SBTN Natural Lands Map |
+    | **Description**<br>A descriptive summary of what the dataset is and key details, such as information on geographic coverage, temporal range, and the data sources used. | The SBTN Natural Lands Map v1.1 is a 2020 baseline map of natural and non-natural land covers intended for use by companies setting science-based targets for nature, specifically the SBTN Land target #1: no conversion of natural ecosystems.<br>This map is global with 30m resolution and was made by compiling existing global and regional data including the GLAD Global Land Cover and Change data, ESA WorldCover, and many other land cover and use datasets.  <br"Natural" and "non-natural" definitions were adapted from the Accountability Framework initiative's definition of a natural ecosystem as "one that substantially resembles - in terms of species composition, structure, and ecological function - what would be found in a given area in the absence of major human impacts" and can include managed ecosystems as well as degraded ecosystems that are expected to regenerate either naturally or through management (AFi 2024). The SBTN Natural Lands Map operationalizes this definition by using proxies based on available data that align with AFi guidance to the extent possible. |
+    | **Function / usage notes in Zeno**<br>The intended use of the dataset. How might users employ this data in order to have some impact in the world?<br>When should Zeno suggest/use this dataset? Please include examples of use cases (e.g. X dataset can be used for XYZ). You can also include these examples in the prompt spreadsheet. | The Natural Lands Map was created for companies to use as a baseline when identifying if there has been conversion of natural ecosystems in their supply chains since 2020. |
+    | **Methodology<br>How was this dataset created? | The SBTN Natural Lands Map is a compilation of many global and regional land cover and land use datasets through a decision tree to disaggregate natural land from non-natural lands, according to definition of natural by the Accountability Framework. Global land cover and use datasets were overlaid with a hierarchy by land cover class, and then regional data replaced the global data, where available. The binary natural/non-natural classes were validated independently by IIASA. |
+    | **Cautions in Zeno**<br>What should be kept in mind when using this dataset?<br>When should Zeno not suggest/use this dataset? Feel free to include examples of when the data should not be used (e.g. X dataset cannot be used for XYZ). | "Natural" and "non-natural" definitions were adapted from the Accountability Framework initiative's definition of a natural ecosystem as "one that substantially resembles - in terms of species composition, structure, and ecological function - what would be found in a given area in the absence of major human impacts" and can include managed ecosystems as well as degraded ecosystems that are expected to regenerate either naturally or through management (AFi 2024). The SBTN Natural Lands Map operationalizes this definition by using proxies based on available data that align with AFi guidance to the extent possible. These definitions may not match all uses of the term “natural” so users should exercise caution when using this map out of the conversion-free supply chains context. <br>This map overestimates the extent of natural lands, and while remote sensing data, on which the map is based, can provide powerful insights, additional field work should be used for validation and to understand local dynamics. Caution should be used if calculating areas with the SBTN Natural Lands Map. |
+    | **Resolution** | 30 x 30 meters |
+    | **Geographic Coverage** | Global |
+    | **Update Frequency** | Irregular |
+    | **Content Date** | 2020 |
+    | **Keywords** | land cover, deforestation, conversion, supply chains, natural lands |
+    | **Band values/meaning**<br>What does the pixel values of each band mean? What are the possible values? | Single band<br>Type: classification<br>binary<br>0: Non-natural<br>1: Natural<br>classification<br>[2-11] natural land classes<br>2: natural forests<br>3: natural short vegetation<br>4: natural water<br>5: mangroves<br>6: bare<br>7: snow<br>8: wetland natural forests<br>9: natural peat forests<br>10: wetland natural short vegetation<br>11: natural peat short vegetation<br>[12-21] non-natural classes<br>12: cropland<br>13: built-up<br>14: non-natural tree cover<br>15: non-natural short vegetation<br>16: non-natural water<br>17: wetland non-natural tree cover <br>18: non-natural peat tree cover<br>19: wetland non-natural short vegetation<br>20: non-natural peat short vegetation<br>21: non-natural bare |
+
+    -----
+
+    **Key Features:**
+    - 🆔 Deterministic UUID generation using SHA-1 hashing (UUIDv5)
+    - 💾 Automatic payload storage in temporary storage
+    - 🔗 Returns a URL to check analytics status
+    - ♻️ Idempotent: Identical payloads return the same resource ID
+
+    **Flow:**
+    1. Accepts `DistAlertsAnalyticsIn` payload
+    2. Generates UUID based on payload content
+    3. Stores payload as JSON file
+    4. Returns resource URL for status checking
+    """
