+++ api/test/unit/test_dist_analytics.py
@@ -2,8 +2,11 @@ import re
 
 import pytest
 
-from api.app.analysis import get_geojson_from_data_api, get_geojson_request_for_data_api
-from api.app.query import create_gadm_dist_query
+from app.analysis.common.analysis import (
+    get_geojson_from_data_api,
+    get_geojson_request_for_data_api,
+)
+from app.analysis.dist_alerts.query import create_gadm_dist_query
 
 
 def test_create_gadm_adm2_dist_query_no_intersection_select_clause():
