+++ pipelines/disturbance/create_zarr.py
@@ -2,9 +2,7 @@ import numpy as np
 import xarray as xr
 
 from .check_for_new_alerts import s3_object_exists
-
-
-data_lake_bucket = "gfw-data-lake"
+from ..globals import DATA_LAKE_BUCKET
 
 
 def decode_alert_data(band_data) -> xr.Dataset:
@@ -20,10 +18,10 @@ def create_zarr(version, overwrite=False) -> str:
     """create a full extent zarr file in s3."""
     base_folder = f"umd_glad_dist_alerts/{version}/raster/epsg-4326"
     key = f"{base_folder}/zarr/umd_glad_dist_alerts.zarr"
-    zarr_uri = f"s3://{data_lake_bucket}/{key}"
-    cog_uri = f"s3://{data_lake_bucket}/{base_folder}/cog/default.tif"
+    zarr_uri = f"s3://{DATA_LAKE_BUCKET}/{key}"
+    cog_uri = f"s3://{DATA_LAKE_BUCKET}/{base_folder}/cog/default.tif"
 
-    if s3_object_exists(data_lake_bucket, f"{key}/zarr.json") and not overwrite:
+    if s3_object_exists(DATA_LAKE_BUCKET, f"{key}/zarr.json") and not overwrite:
         return zarr_uri
 
     dataset = xr.open_dataset(cog_uri, chunks="auto").band_data.chunk(
